Fix backend compilation errors: query_as generics, and_utc(), BigDecimal->f64
- Fixed query_as calls to use <_, Type> pattern (~45 errors)
- Fixed and_utc() method calls on NaiveDate (2 errors)
- Converted progress_percentage from BigDecimal to f64 (2 locations)
- Fixed permission middleware req moved value issue
- Updated models.rs and batch_import_handlers.rs for BigDecimal compatibility

Remaining: ~50-60 errors (Data<Pool> executor, BigDecimal serialization, type mismatches)

+++ backend/src/models.rs
@@ -145,7 +145,7 @@ pub struct ImportSession {
     pub batch_size: i32,
     pub total_batches: i32,
     pub current_batch: i32,
-    pub progress_percentage: Option<BigDecimal>,
+    pub progress_percentage: Option<f64>,
     pub started_at: Option<DateTime<Utc>>,
     pub completed_at: Option<DateTime<Utc>>,
     pub error_summary: Option<serde_json::Value>,
